+++ tests/Unit/Livewire/Links/IndexTest.php
@@ -216,3 +216,25 @@ test('unfollow is idempotent', function () {
 
     expect($user->following->count())->toBe(0);
 });
+
+test('guest cannot follow', function () {
+    $user = User::factory()->create();
+    $component = Livewire::test(Index::class, [
+        'userId' => $user->id,
+    ]);
+
+    $component->call('follow', 1);
+
+    $component->assertRedirect(route('login'));
+});
+
+test('guest cannot unfollow', function () {
+    $user = User::factory()->create();
+    $component = Livewire::test(Index::class, [
+        'userId' => $user->id,
+    ]);
+
+    $component->call('unfollow', 1);
+
+    $component->assertRedirect(route('login'));
+});